# [perf] split out pretty and create hot paths