--- a/.changeset/silly-masks-jump.md
+++ b/.changeset/silly-masks-jump.md
@@ -1,0 +1,5 @@
+---
+'preact-render-to-string': patch
+---
+
+Split up hot paths and make separate path for opts.pretty