This PR aims to create as many hot paths as possible, we went from being slower than react 18 string rendering to

## FROM

```
Running "search-results"...

Running benchmark "preact"...
preact x 1,003 ops/sec ±1.53% (88 runs sampled)

Running benchmark "react"...
react x 1,127 ops/sec ±6.13% (76 runs sampled)

--------------
Running "color-picker"...

Running benchmark "preact"...
preact x 5,513 ops/sec ±0.29% (93 runs sampled)

Running benchmark "react"...
react x 5,829 ops/sec ±8.04% (79 runs sampled)
--------------
```

## TO

```
Running "search-results"...

Running benchmark "preact"...
preact x 1,437 ops/sec ±1.56% (93 runs sampled)

Running benchmark "react"...
react x 1,235 ops/sec ±5.94% (82 runs sampled)

--------------

Running "color-picker"...

Running benchmark "preact"...
preact x 7,862 ops/sec ±0.31% (92 runs sampled)

Running benchmark "react"...
react x 6,083 ops/sec ±8.30% (78 runs sampled)

--------------
```